+++ src/main/java/src/main/java/duke/Deadline.java
@@ -1,3 +1,5 @@
+package src.main.java.duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
