+++ src/main/java/src/main/java/duke/DukeException.java
@@ -1,3 +1,5 @@
+package src.main.java.duke;
+
 public class DukeException extends Exception {
 
     DukeException(String err) {
