+++ src/main/java/src/main/java/duke/InvalidInputException.java
@@ -1,3 +1,5 @@
+package src.main.java.duke;
+
 public class InvalidInputException extends DukeException {
 
     InvalidInputException(String err) {
