+++ src/main/java/src/main/java/duke/NoDescriptionException.java
@@ -1,3 +1,5 @@
+package src.main.java.duke;
+
 public class NoDescriptionException extends DukeException {
 
     NoDescriptionException(String err) {
