+++ src/main/java/src/main/java/duke/NoNumberException.java
@@ -1,3 +1,5 @@
+package src.main.java.duke;
+
 public class NoNumberException extends DukeException {
 
     NoNumberException(String err) {
