+++ src/main/java/src/main/java/duke/TaskNotFoundException.java
@@ -1,3 +1,5 @@
+package src.main.java.duke;
+
 public class TaskNotFoundException extends DukeException {
 
     TaskNotFoundException(String err) {
